requirements.txt
Unidecode==1.0.22
ckantoolkit==0.0.3
iribaker==0.1.2
pandas==0.24.2
pysolr==3.6.0
python-dateutil==2.8.1
pyyaml==5.4
requests[security]==2.25.1
urllib3[secure]==1.26.5
isodate==0.6.1
